move runtype to OSS
