+++ enterprise/dev/ci/internal/ci/pipeline.go
@@ -14,13 +14,13 @@ import (
 	"github.com/google/go-github/v41/github"
 	"github.com/slack-go/slack"
 
+	"github.com/sourcegraph/sourcegraph/dev/ci/runtype"
 	"github.com/sourcegraph/sourcegraph/dev/team"
 	"github.com/sourcegraph/sourcegraph/enterprise/dev/ci/images"
 	"github.com/sourcegraph/sourcegraph/enterprise/dev/ci/internal/buildkite"
 	bk "github.com/sourcegraph/sourcegraph/enterprise/dev/ci/internal/buildkite"
 	"github.com/sourcegraph/sourcegraph/enterprise/dev/ci/internal/ci/changed"
 	"github.com/sourcegraph/sourcegraph/enterprise/dev/ci/internal/ci/operations"
-	"github.com/sourcegraph/sourcegraph/enterprise/dev/ci/runtype"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
